+++ src/argument_processors/argument_processor_config.h
@@ -27,4 +27,22 @@ class ArgumentProcessorConfigTestEmail : public WsjcppArgumentProcessor {
         virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
 };
 
+class ArgumentProcessorConfigLxdEnable : public WsjcppArgumentProcessor {
+    public:
+        ArgumentProcessorConfigLxdEnable();
+        virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
+};
+
+class ArgumentProcessorConfigLxdDisable : public WsjcppArgumentProcessor {
+    public:
+        ArgumentProcessorConfigLxdDisable();
+        virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
+};
+
+class ArgumentProcessorConfigLxdConfigure : public WsjcppArgumentProcessor {
+    public:
+        ArgumentProcessorConfigLxdConfigure();
+        virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
+};
+
 #endif // ARGUMENT_PROCESSOR_CONFIG_H